Stop crash when one part of mosaic doesn't exist
Currently, the mosaicking in postprocessing will skip stitching if any "expected" files don't exist. But, it's not very smart about checking these things (it expects every part to have every configuration), so you can have an instance where you have mosaic parts that might not have certain configurations:

- ngc7793_1 only has 12m
- ngc7793_2 has 7m
- ngc7793_3 has 7m
- ngc7793_4 has 12m+7m+tp

and then it'll not do a 12m mosaic of 1 and 4, or a 7m mosaic of 2/3/4. This adds some logic to check parts exist first, thus avoiding the skip. Also adds warnings if files are missing, in case this isn't intended behaviour for the end user.

Importantly, this works if the heterogeneous observations don't have overlap between each other. It then just saves duplicating targets to avoid this behaviour